demo2/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0
torchvision>=0.15.0

# Feature extraction
angr>=9.2.0
capstone>=5.0.0
pefile>=2023.2.7
yara-python>=4.3.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Note: Ghidra API requires Java and Ghidra installation
# See: https://ghidra-sre.org/InstallationGuide.html

